Move tests from Multinetwork-test classes to base classes in preparation for merging the Multinetwork-classes with the mutable classes.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=305434560
