+++ guava-tests/test/com/google/common/graph/AbstractStandardDirectedNetworkTest.java
@@ -525,8 +525,9 @@ public abstract class AbstractStandardDirectedNetworkTest extends AbstractNetwor
   }
 
   @Test
-  public void addEdge_parallelEdge() {
+  public void addEdge_parallelEdge_notAllowed() {
     assume().that(graphIsMutable()).isTrue();
+    assume().that(network.allowsParallelEdges()).isFalse();
 
     addEdge(N1, N2, E12);
     try {
@@ -535,6 +536,16 @@ public abstract class AbstractStandardDirectedNetworkTest extends AbstractNetwor
     } catch (IllegalArgumentException e) {
       assertThat(e).hasMessageThat().contains(ERROR_PARALLEL_EDGE);
     }
+  }
+
+  @Test
+  public void addEdge_parallelEdge_allowsParallelEdges() {
+    assume().that(graphIsMutable()).isTrue();
+    assume().that(network.allowsParallelEdges()).isTrue();
+
+    assertTrue(networkAsMutableNetwork.addEdge(N1, N2, E12));
+    assertTrue(networkAsMutableNetwork.addEdge(N1, N2, E12_A));
+    assertThat(network.edgesConnecting(N1, N2)).containsExactly(E12, E12_A);
   }
 
   @Test
@@ -635,9 +646,10 @@ public abstract class AbstractStandardDirectedNetworkTest extends AbstractNetwor
   }
 
   @Test
-  public void addEdge_parallelSelfLoopEdge() {
+  public void addEdge_parallelSelfLoopEdge_notAllowed() {
     assume().that(graphIsMutable()).isTrue();
     assume().that(network.allowsSelfLoops()).isTrue();
+    assume().that(network.allowsParallelEdges()).isFalse();
 
     addEdge(N1, N1, E11);
     try {
@@ -646,6 +658,17 @@ public abstract class AbstractStandardDirectedNetworkTest extends AbstractNetwor
     } catch (IllegalArgumentException e) {
       assertThat(e.getMessage()).contains(ERROR_PARALLEL_EDGE);
     }
+  }
+
+  @Test
+  public void addEdge_parallelSelfLoopEdge_allowsParallelEdges() {
+    assume().that(graphIsMutable()).isTrue();
+    assume().that(network.allowsSelfLoops()).isTrue();
+    assume().that(network.allowsParallelEdges()).isTrue();
+
+    assertTrue(networkAsMutableNetwork.addEdge(N1, N1, E11));
+    assertTrue(networkAsMutableNetwork.addEdge(N1, N1, E11_A));
+    assertThat(network.edgesConnecting(N1, N1)).containsExactly(E11, E11_A);
   }
 
   @Test
