+++ guava-tests/test/com/google/common/graph/StandardDirectedMultiNetworkTest.java
@@ -16,11 +16,6 @@
 
 package com.google.common.graph;
 
-import static com.google.common.truth.Truth.assertThat;
-import static com.google.common.truth.TruthJUnit.assume;
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
@@ -41,78 +36,5 @@ public class StandardDirectedMultiNetworkTest extends AbstractStandardDirectedNe
   @Override
   void addEdge(Integer n1, Integer n2, String e) {
     networkAsMutableNetwork.addEdge(n1, n2, e);
-  }
-
-  @Test
-  public void adjacentEdges_parallelEdges() {
-    addEdge(N1, N2, E12);
-    addEdge(N1, N2, E12_A);
-    addEdge(N1, N2, E12_B);
-    addEdge(N3, N4, E34);
-    assertThat(network.adjacentEdges(E12)).containsExactly(E12_A, E12_B);
-  }
-
-  @Test
-  public void edgesConnecting_parallelEdges() {
-    addEdge(N1, N2, E12);
-    addEdge(N1, N2, E12_A);
-
-    assertThat(network.edgesConnecting(N1, N2)).containsExactly(E12, E12_A);
-    // Passed nodes should be in the correct edge direction, first is the
-    // source node and the second is the target node
-    assertThat(network.edgesConnecting(N2, N1)).isEmpty();
-  }
-
-  @Test
-  public void edgesConnecting_parallelSelfLoopEdges() {
-    addEdge(N1, N1, E11);
-    addEdge(N1, N1, E11_A);
-
-    assertThat(network.edgesConnecting(N1, N1)).containsExactly(E11, E11_A);
-  }
-
-  @Override
-  @Test
-  public void addEdge_parallelEdge() {
-    assume().that(graphIsMutable()).isTrue();
-
-    assertTrue(networkAsMutableNetwork.addEdge(N1, N2, E12));
-    assertTrue(networkAsMutableNetwork.addEdge(N1, N2, E12_A));
-    assertThat(network.edgesConnecting(N1, N2)).containsExactly(E12, E12_A);
-  }
-
-  @Override
-  @Test
-  public void addEdge_parallelSelfLoopEdge() {
-    assume().that(graphIsMutable()).isTrue();
-
-    assertTrue(networkAsMutableNetwork.addEdge(N1, N1, E11));
-    assertTrue(networkAsMutableNetwork.addEdge(N1, N1, E11_A));
-    assertThat(network.edgesConnecting(N1, N1)).containsExactly(E11, E11_A);
-  }
-
-  @Test
-  public void removeEdge_parallelEdge() {
-    assume().that(graphIsMutable()).isTrue();
-
-    addEdge(N1, N2, E12);
-    addEdge(N1, N2, E12_A);
-    assertTrue(networkAsMutableNetwork.removeEdge(E12_A));
-    assertThat(network.edgesConnecting(N1, N2)).containsExactly(E12);
-  }
-
-  @Test
-  public void removeEdge_parallelSelfLoopEdge() {
-    assume().that(graphIsMutable()).isTrue();
-
-    addEdge(N1, N1, E11);
-    addEdge(N1, N1, E11_A);
-    addEdge(N1, N2, E12);
-    assertTrue(networkAsMutableNetwork.removeEdge(E11_A));
-    assertThat(network.edgesConnecting(N1, N1)).containsExactly(E11);
-    assertThat(network.edgesConnecting(N1, N2)).containsExactly(E12);
-    assertTrue(networkAsMutableNetwork.removeEdge(E11));
-    assertThat(network.edgesConnecting(N1, N1)).isEmpty();
-    assertThat(network.edgesConnecting(N1, N2)).containsExactly(E12);
   }
 }
